% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anl_align.R
\name{anl_align}
\alias{anl_align}
\title{window}
\usage{
anl_align(
  h,
  gselbtn = svalue(gsel_button),
  dir = svalue(fq_dir),
  refP1 = svalue(ref_P1),
  ...
)
}
\value{
anl_align
}
\description{
window
}
