% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anl_filter.R
\name{anl_filter}
\alias{anl_filter}
\title{window}
\usage{
anl_filter(
  h,
  selrcbtn = svalue(selrc_button),
  cfilv = svalue(cfilv),
  sfil = svalue(sfil),
  ...
)
}
\value{
anl_filter
}
\description{
window
}
