% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_lengths}
\alias{check_lengths}
\title{Check vector argument lengths}
\usage{
check_lengths(...)
}
\arguments{
\item{...}{Numeric vectors to check.}
}
\value{
Invisible `NULL`. An error is thrown if lengths are incompatible.
}
\description{
Ensure that all supplied arguments are either scalars or vectors of the same length.
}
\keyword{internal}
