% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{commodity_al_doses}
\alias{commodity_al_doses}
\title{Estimate total number of AL doses required}
\usage{
commodity_al_doses(n_cases, treatment_coverage, proportion_act, age_upper)
}
\arguments{
\item{n_cases}{Vector of malaria case numbers by age band.}

\item{treatment_coverage}{Vector of treatment coverage proportions by age band.}

\item{proportion_act}{Vector of proportion of treatments that are ACTs by age band.}

\item{age_upper}{Vector of upper bounds for each age group.}
}
\value{
A vector giving the number of 20/120mg Artemether + lumefantrine ACT doses required per age group.
}
\description{
Note the cost per dose is for a single dose (20/120 mg). A treatment course typically
constitutes Artemether + lumefantrine given twice a day for 3 days following
weight-based guidelines:
\itemize{
  \item 5 to <15 kg: 20/120 mg
  \item 15 to <25 kg: 40/240 mg
  \item 25 to <35 kg: 60/360 mg
  \item >=35 kg: 80/480 mg
}
So course for a single adult (weighing >=35kg) may constitute
3 days x 2 times daily x 4 doses (4 x 20/120mg = 80/480mg) = 24 doses.
}
