% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc.R
\name{commodity_doses_smc}
\alias{commodity_doses_smc}
\title{Number of doses of SMC}
\usage{
commodity_doses_smc(smc_cov, n_rounds, par_smc)
}
\arguments{
\item{smc_cov}{SMC coverage. Numeric scalar or vector.}

\item{n_rounds}{Number of SMC rounds per year. Numeric scalar or vector.}

\item{par_smc}{Population at risk within SMC-eligible age range. Numeric scalar or vector.}
}
\value{
The total number of SMC doses delivered.
}
\description{
Number of doses of SMC
}
