% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs.R
\name{commodity_households_irs}
\alias{commodity_households_irs}
\title{Number of people protected by IRS}
\usage{
commodity_households_irs(irs_cov, n_rounds, par, hh_size)
}
\arguments{
\item{irs_cov}{A single value or vector of IRS coverage.}

\item{n_rounds}{The number of spray rounds per year}

\item{par}{Population at risk estimates.}

\item{hh_size}{The average number of occupants per household}
}
\value{
The total number of household-rounds of IRS protection.
}
\description{
Number of people protected by IRS
}
