% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{commodity_microscopy_tests}
\alias{commodity_microscopy_tests}
\title{Estimate total number of microscopy diagnostics required}
\usage{
commodity_microscopy_tests(
  n_cases,
  treatment_coverage,
  proportion_microscopy,
  proportion_tested = 1
)
}
\arguments{
\item{n_cases}{Malaria case numbers. Numeric scalar or vector.}

\item{treatment_coverage}{Treatment coverage. Numeric scalar or vector.}

\item{proportion_microscopy}{Proportion of diagnostics that are microscopy. Numeric scalar or vector.}

\item{proportion_tested}{Proportion of treated cases that are tested. Numeric scalar or vector.}
}
\value{
Vector of the number of microscopy tests
}
\description{
Estimate total number of microscopy diagnostics required
}
