% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{commodity_nmf_rdt_tests}
\alias{commodity_nmf_rdt_tests}
\title{Estimate total number of RDT diagnostics required as a result of non malarial fevers}
\usage{
commodity_nmf_rdt_tests(
  n_nmf,
  treatment_coverage,
  proportion_rdt,
  proportion_tested = 1,
  pfpr,
  pfpr_threshold = 0.05
)
}
\arguments{
\item{n_nmf}{Non malarial fever case numbers. Numeric scalar or vector.}

\item{treatment_coverage}{Treatment coverage. Numeric scalar or vector.}

\item{proportion_rdt}{Proportion of diagnostics that are RDT. Numeric scalar or vector.}

\item{proportion_tested}{Proportion of NMFs that are tested. Numeric scalar or vector.}

\item{pfpr}{Prevalence. Numeric scalar or vector.}

\item{pfpr_threshold}{Prevalence threshold at which it is assumed NMF are not suspected (and subsequently tested) to be malaria. Numeric scalar or vector.}
}
\value{
Vector of the number of RDTs tests used on NMFs
}
\description{
Estimate total number of RDT diagnostics required as a result of non malarial fevers
}
