% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs.R
\name{commodity_persons_irs}
\alias{commodity_persons_irs}
\title{Number of people protected by IRS}
\usage{
commodity_persons_irs(irs_cov, n_rounds, par)
}
\arguments{
\item{irs_cov}{A single value or vector of IRS coverage.}

\item{n_rounds}{The number of spray rounds per year}

\item{par}{Population at risk estimates.}
}
\value{
The total number of person-rounds of IRS protection.
}
\description{
Number of people protected by IRS
}
