% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{cost_al}
\alias{cost_al}
\title{Cost Artemether/Lumefantrine treatment}
\usage{
cost_al(n_doses, cost_per_dose = 0.3, input_year = 2022, ...)
}
\arguments{
\item{n_doses}{Number of doses}

\item{cost_per_dose}{Cost per dose is for a single dose (20/120 mg)}

\item{input_year}{Year the unit costs are reported in}

\item{...}{Additional arguments passed to `inflation_adjust()`}
}
\value{
AL costs
}
\description{
Note the cost per dose is for a single dose (20/120 mg), not a full treatment course.
}
\references{
\strong{dosing information}

\url{https://www.who.int/publications/i/item/guidelines-for-malaria}

\strong{cost_per_dose}

Current default is the average cost for 20/120mg.

The Global Fund Pooled Procurement Mechanism Reference Pricing: Antimalarial medicines, version: quarter 1, 2022

\url{https://www.theglobalfund.org/en/sourcing-management/health-products/antimalarial-medicines/}.
}
