% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs.R
\name{cost_ddt_irs_structure}
\alias{cost_ddt_irs_structure}
\title{Cost DDT IRS}
\usage{
cost_ddt_irs_structure(
  n_sprayed,
  cost_per_structure_sprayed = 2.25,
  input_year = 1999,
  ...
)
}
\arguments{
\item{n_sprayed}{Number of structures sprayed}

\item{cost_per_structure_sprayed}{Cost per structure sprayed}

\item{input_year}{Year the unit costs are reported in}

\item{...}{Additional arguments passed to `inflation_adjust()`}
}
\value{
DDT IRS costs
}
\description{
Cost DDT IRS
}
\references{
\strong{cost_per_structure_sprayed}

Current default is the (mid point) average cost for DDT IRS
from Table 3 of Walker (2008)

\url{https://resjournals.onlinelibrary.wiley.com/doi/10.1046/j.1365-2915.2000.00262.x}.
}
