% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipti.R
\name{cost_ipti}
\alias{cost_ipti}
\title{Cost IPTi}
\usage{
cost_ipti(n_doses, ipti_cost_per_dose_delivered = 0.1298)
}
\arguments{
\item{n_doses}{Number of IPTi doses}

\item{ipti_cost_per_dose_delivered}{Cost per dose delivered}
}
\value{
IPTi costs
}
\description{
Cost IPTi
}
\references{
\strong{IPTi_cost_per_dose_delivered}

Current default is the average economic cost of administering 3 rounds of IPTi
  (annually), of sulfadoxine–pyrimethamine in Tanzania, Ghana, Mozambique and Gabon.
   The cost per dose delivered is the average cost of trial results for
   three IPTi cycles divided by the number of cycles (0.39 / 3). Costs
   are in 2007 USD$.

Conteh et al (2010) table S4

\url{https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0010313}.
}
