% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs.R
\name{cost_ll_irs_person}
\alias{cost_ll_irs_person}
\title{Cost long lasting IRS}
\usage{
cost_ll_irs_person(
  n_protected,
  cost_per_person_protected = 7.44,
  input_year = 2020,
  ...
)
}
\arguments{
\item{n_protected}{Number of people protected}

\item{cost_per_person_protected}{Cost per person protected}

\item{input_year}{Year the unit costs are reported in}

\item{...}{Additional arguments passed to `inflation_adjust()`}
}
\value{
Long lasting IRS costs
}
\description{
Cost long lasting IRS
}
\references{
\strong{cost_per_person_protected}

Current default is the (unweighted) average economic cost for long lasting
(Oganophosphate/Neonicotinoid/Neonicotinoid/pyrethroid mix) IRS
from PMI reports for: Uganda, Zambia, Ethiopia, Tanzania, Kenya, Mozambique,
Benin, Ghana, Rwanda, Madagascar, Burkina Faso, Senegal, Zimbabwe, Mali
Malawi and Cote d'Ivoire

PMI IRS Country Programs: 2020, Comparative Cost Analysis, table CC2

\url{https://www.pmi.gov/pmi-vectorlink-cost-study-report_2020_approved-june-14-2021-sxf-508/}.
}
