% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bet_nets.R
\name{cost_llin}
\alias{cost_llin}
\title{Cost standard LLINS}
\usage{
cost_llin(
  n_llin,
  llin_unit_cost = 2.02,
  llin_delivery_cost = 1.5,
  input_year = 2024,
  ...
)
}
\arguments{
\item{n_llin}{Number of standard LLIN bed nets}

\item{llin_unit_cost}{Commodity unit cost per standard LLIN bed net.}

\item{llin_delivery_cost}{Cost to deliver one standard LLIN bet net.}

\item{input_year}{Year the unit costs are reported in}

\item{...}{Additional arguments passed to `inflation_adjust()`}
}
\value{
LLIN costs
}
\description{
Cost standard LLINS
}
\references{
\strong{llin_unit_cost}

Current default is the average cost for a pyrethroid-only ITN, including,
  hooks, strings, bag and customisation.

The Global Fund Pooled Procurement Mechanism Reference Pricing: Insecticide-Treated Nets, accessed 16-12-2024

\url{https://www.theglobalfund.org/en/sourcing-management/health-products/long-lasting-insecticidal-nets/}.

\strong{llin_delivery_cost}

Sherrard-Smith et al (2022)

\url{https://www.thelancet.com/journals/lanplh/article/PIIS2542-5196(21)00296-5/fulltext}.
}
