% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{cost_outpatient}
\alias{cost_outpatient}
\title{Cost per outpatient visit}
\usage{
cost_outpatient(n_visits, cost_per_visit, input_year = 2021, ...)
}
\arguments{
\item{n_visits}{Number of visits. Numeric scalar or vector.}

\item{cost_per_visit}{Cost per visit. Numeric scalar or vector.}

\item{input_year}{Year the unit costs are reported in}

\item{...}{Additional arguments passed to `inflation_adjust()`}
}
\value{
Outpatient costs
}
\description{
For information on country specific outpatient costs from WHO CHOICE see \code{?who_coice}
}
