% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bet_nets.R
\name{cost_pbo_itn}
\alias{cost_pbo_itn}
\title{Cost pyrethroid-PBO ITN}
\usage{
cost_pbo_itn(
  n_pbo_itn,
  pbo_itn_unit_cost = 2.63,
  pbo_itn_delivery_cost = 1.5,
  input_year = 2024,
  ...
)
}
\arguments{
\item{n_pbo_itn}{Number of pyrethroid-PBO bed nets. Numeric scalar or vector.}

\item{pbo_itn_unit_cost}{Commodity unit cost per pyrethroid-PBO ITN bed net. Numeric scalar or vector.}

\item{pbo_itn_delivery_cost}{Cost to deliver one pyrethroid-PBO ITN bet net. Numeric scalar or vector.}

\item{input_year}{Year the unit costs are reported in}

\item{...}{Additional arguments passed to `inflation_adjust()`}
}
\value{
LLIN costs
}
\description{
Cost pyrethroid-PBO ITN
}
\references{
\strong{pbo_itn_unit_cost}

Current default is the average cost for a pyrethroid-PBO ITN, including,
  hooks, strings, bag and customisation.

The Global Fund Pooled Procurement Mechanism Reference Pricing: Insecticide-Treated Nets, accessed 16-12-2024

\url{https://www.theglobalfund.org/en/sourcing-management/health-products/long-lasting-insecticidal-nets/}.

\strong{pbo_itn_delivery_cost}

Sherrard-Smith et al (2022)

\url{https://www.thelancet.com/journals/lanplh/article/PIIS2542-5196(21)00296-5/fulltext}.
}
