% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance.R
\name{cost_surveillance}
\alias{cost_surveillance}
\title{Cost surveillance}
\usage{
cost_surveillance(
  pop_at_risk,
  cost_per_pop_at_risk = 0.05,
  input_year = 2017,
  ...
)
}
\arguments{
\item{pop_at_risk}{Population at risk. Numeric scalar or vector.}

\item{cost_per_pop_at_risk}{Cost per population at risk. Numeric scalar or vector.}

\item{input_year}{Year the unit costs are reported in}

\item{...}{Additional arguments passed to `inflation_adjust()`}
}
\value{
Surveillance costs
}
\description{
Cost of epidemiological and entomological surveillance.
}
\references{
\strong{cost_per_pop_at_risk}

Estimate from Patouillard et al (2017), consisting of a summary from the strategic
plans from: Botswana, Nigeria, India, Eritrea, Swaziland, Namibia.

\url{https://gh.bmj.com/content/2/2/e000176}.
}
