% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_options.R
\name{set_region}
\alias{set_region}
\title{Set the global region for inflation adjustments}
\usage{
set_region(region)
}
\arguments{
\item{region}{Target region to use when adjusting for inflation}
}
\description{
Set the global region for inflation adjustments
}
