% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{who_choice}
\alias{who_choice}
\title{WHO CHOICE inpatient/outpatient costs.}
\format{
A data frame with 194 rows and 5 variables:
\describe{
  \item{ISO}{Country iso3c code}
  \item{world_bank_income_group}{World bank income group}
  \item{world_bank_region}{World bank region}
  \item{inpatient}{Inpatient visit cost}
  \item{outpatient}{Outpatient visit cost}
}
}
\source{
\url{http://www.who.int/choice/en/}
}
\usage{
who_choice
}
\description{
Data from: World Health Organization.
  Cost effectiveness and strategic planning (WHO-CHOICE)
}
\keyword{datasets}
