% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{commodity_chloroquine_doses}
\alias{commodity_chloroquine_doses}
\title{Estimate total number of Chloroquine doses required}
\usage{
commodity_chloroquine_doses(
  n_cases,
  treatment_coverage,
  proportion_non_act,
  age_upper
)
}
\arguments{
\item{n_cases}{Vector of malaria case numbers by age band.}

\item{treatment_coverage}{Vector of treatment coverage proportions.}

\item{proportion_non_act}{Vector of proportion of treatments that are non-ACT (assumed chloroquine).}

\item{age_upper}{Vector of upper bounds for each age group.}
}
\value{
A vector giving the number of 250mg chloroquine doses required per age group.
}
\description{
Note the cost per dose is for a single dose (250 mg chloroquine phosphate).
A treatment course typically constitutes 41mg/kg (max 2500mg) for adults
\itemize{
  \item 5 to <15 kg: 2 doses
  \item 15 to <25 kg: 4 doses
  \item 25 to <35 kg: 5 doses
  \item >=35 kg: 10 doses
}
So course for a single adult may constitute approximately 10 doses.
}
