% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmc.R
\name{commodity_doses_pmc}
\alias{commodity_doses_pmc}
\title{Number of doses of PMC}
\usage{
commodity_doses_pmc(pmc_cov, par_pmc, n_rounds = 3)
}
\arguments{
\item{pmc_cov}{A single value or vector of pmc coverage.}

\item{par_pmc}{Population at risk within pmc-eligable age range estimates.}

\item{n_rounds}{The number of pmc rounds per year}
}
\value{
The total number of pmc doses delivered.
}
\description{
Number of doses of PMC
}
