% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccine.R
\name{commodity_doses_vaccine}
\alias{commodity_doses_vaccine}
\title{Number of vaccine doses delivered}
\usage{
commodity_doses_vaccine(
  vaccine_cov,
  par_vaccine,
  n_dose_primary_series = 3,
  booster_coverage_downscale = 0.8,
  n_boosters = 1
)
}
\arguments{
\item{vaccine_cov}{A single value or vector of vaccine coverage.}

\item{par_vaccine}{Population at risk within vaccine-eligible age range estimates.}

\item{n_dose_primary_series}{Number of doses in the primary series}

\item{booster_coverage_downscale}{Drop off in coverage between primary series and
boosters such that `booster coverage = vaccine_cov * booster_coverage_downscale`}

\item{n_boosters}{Number of booster doses}
}
\value{
The total number of vaccine doses delivered.
}
\description{
Note, if duration for full course (primary + boosters) is >1 year, costs for
the full course will be all assigned to the year of the primary series.
}
