% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{commodity_microscopy_tests}
\alias{commodity_microscopy_tests}
\title{Estimate total number of microscopy diagnostics required}
\usage{
commodity_microscopy_tests(
  n_cases,
  treatment_coverage,
  proportion_microscopy,
  proportion_tested = 1
)
}
\arguments{
\item{n_cases}{Vector of malaria case numbers}

\item{treatment_coverage}{Vector of treatment coverage}

\item{proportion_microscopy}{Vector of proportion of diagnostics that are microscopy}

\item{proportion_tested}{A scalar value of proportion of treated cases that are tested}
}
\value{
Vector of the number of microscopy tests
}
\description{
Estimate total number of microscopy diagnostics required
}
