% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bet_nets.R
\name{commodity_nets}
\alias{commodity_nets}
\title{Estimate the number of bed nets required to match usage target}
\usage{
commodity_nets(
  usage,
  use_rate,
  distribution_timesteps,
  crop_timesteps,
  half_life,
  par,
  ...
)
}
\arguments{
\item{usage}{A single value or vector of desired target usages (%) to model.}

\item{use_rate}{A single value or vector of usage rates.}

\item{distribution_timesteps}{Timesteps of distributions (days). By default,
we can assume that net distributions happen on the first day of each year.
For example c(1, 366)}

\item{crop_timesteps}{Timesteps of crop estimates (days). If assuming distribtions
occur on the first day of each year, a reasonable assumption would be that the
crop (and therefore corresponding usage) estimates were taken at the mid-point of each year.
For example c(1, 366) + 183.}

\item{half_life}{Country-specific half-life of nets in days.}

\item{par}{Population at risk estimates. A vector the same length as usage}

\item{...}{additional arguments for the crop_to_distribution function in netz}
}
\value{
Number of nets required to match usage targets
}
\description{
Estimate the number of bed nets required to match usage target
}
\references{
reference
Uses a version of the net stock and flow model as described by:
Bertozzi-Villa, Amelia, et al. Nature communications 12.1 (2021): 1-12.
}
