% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bet_nets.R
\name{commodity_nets}
\alias{commodity_nets}
\title{Estimate the number of bed nets required to match usage target}
\usage{
commodity_nets(
  usage,
  use_rate,
  distribution_timesteps,
  crop_timesteps,
  half_life,
  par,
  ...
)
}
\arguments{
\item{usage}{Desired target usages (\%) to model. Numeric scalar or vector.}

\item{use_rate}{Usage rates. Numeric scalar or vector.}

\item{distribution_timesteps}{Distribution timesteps (days). Numeric scalar or vector. By default,
we can assume that net distributions happen on the first day of each year.
For example c(1, 366)}

\item{crop_timesteps}{Crop estimate timesteps (days). Numeric scalar or vector. If assuming distributions
occur on the first day of each year, a reasonable assumption would be that the
crop (and therefore corresponding usage) estimates were taken at the mid-point of each year.
For example c(1, 366) + 183.}

\item{half_life}{Country-specific net half-life in days. Numeric scalar or vector.}

\item{par}{Population at risk estimates. Numeric scalar or vector.}

\item{...}{additional arguments for the crop_to_distribution function in netz}
}
\value{
Number of nets required to match usage targets
}
\description{
Estimate the number of bed nets required to match usage target
}
\references{
reference
Uses a version of the net stock and flow model as described by:
Bertozzi-Villa, Amelia, et al. Nature communications 12.1 (2021): 1-12.
}
