% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{commodity_nmf_microscopy_tests}
\alias{commodity_nmf_microscopy_tests}
\title{Estimate total number of microscopy diagnostics required as a result of non malarial fevers}
\usage{
commodity_nmf_microscopy_tests(
  n_nmf,
  treatment_coverage,
  proportion_microscopy,
  proportion_tested = 1,
  pfpr,
  pfpr_threshold = 0.05
)
}
\arguments{
\item{proportion_microscopy}{Proportion of diagnostics that are microscopy}

\item{inheritparams}{commodity_nmf_rdt_tests}
}
\description{
Estimate total number of microscopy diagnostics required as a result of non malarial fevers
}
