% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{commodity_nmf_rdt_tests}
\alias{commodity_nmf_rdt_tests}
\title{Estimate total number of RDT diagnostics required as a result of non malarial fevers}
\usage{
commodity_nmf_rdt_tests(
  n_nmf,
  treatment_coverage,
  proportion_rdt,
  proportion_tested = 1,
  pfpr,
  pfpr_threshold = 0.05
)
}
\arguments{
\item{n_nmf}{Vector of non malarial fever case numbers}

\item{treatment_coverage}{Treatment coverage}

\item{proportion_rdt}{Proportion of diagnostics that are RDT}

\item{proportion_tested}{Proportion of nmfs that are tested}

\item{pfpr}{Prevalence}

\item{pfpr_threshold}{Prevalence threshold at which it is assummed NMF are not suspected (and subsequently tested) to be malaria}
}
\description{
Estimate total number of RDT diagnostics required as a result of non malarial fevers
}
