% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs.R
\name{commodity_person_rounds_irs}
\alias{commodity_person_rounds_irs}
\title{Number of people-rounds of IRS}
\usage{
commodity_person_rounds_irs(irs_cov, n_rounds, par)
}
\arguments{
\item{irs_cov}{A single value or vector of IRS coverage.}

\item{n_rounds}{The number of spray rounds per year}

\item{par}{Population at risk estimates.}
}
\value{
The total number of person-rounds of IRS protection.
}
\description{
Number of people-rounds of IRS
}
