% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{commodity_rdt_tests}
\alias{commodity_rdt_tests}
\title{Estimate total number of RDT diagnostics required}
\usage{
commodity_rdt_tests(
  n_cases,
  treatment_coverage,
  proportion_rdt,
  proportion_tested = 1
)
}
\arguments{
\item{n_cases}{Malaria case numbers. Numeric scalar or vector.}

\item{treatment_coverage}{Treatment coverage. Numeric scalar or vector.}

\item{proportion_rdt}{Proportion of diagnostics that are RDT. Numeric scalar or vector.}

\item{proportion_tested}{Proportion of treated cases that are tested. Numeric scalar or vector.}
}
\value{
Vector of the number of RDT tests
}
\description{
Estimate total number of RDT diagnostics required
}
