% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs.R
\name{commodity_structure_rounds_irs}
\alias{commodity_structure_rounds_irs}
\title{Number of structure-rounds of IRS}
\usage{
commodity_structure_rounds_irs(irs_cov, n_rounds, par, hh_size)
}
\arguments{
\item{irs_cov}{IRS coverage per year. Numeric scalar or vector.}

\item{n_rounds}{Number of spray rounds per year. Numeric scalar or vector.}

\item{par}{Population at risk estimates. Numeric scalar or vector.}

\item{hh_size}{Average number of occupants per household. Numeric scalar or vector.}
}
\value{
The total number of structure-rounds of IRS protection.
}
\description{
Assumes 1 structure per household
}
