% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs.R
\name{commodity_structure_rounds_irs}
\alias{commodity_structure_rounds_irs}
\title{Number of structure-rounds of IRS}
\usage{
commodity_structure_rounds_irs(irs_cov, n_rounds, par, hh_size)
}
\arguments{
\item{irs_cov}{A single value or vector of IRS coverage.}

\item{n_rounds}{The number of spray rounds per year}

\item{par}{Population at risk estimates.}

\item{hh_size}{The average number of occupants per household}
}
\value{
The total number of structure-rounds of IRS protection.
}
\description{
Assumes 1 structure per household
}
