% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{cost_al}
\alias{cost_al}
\title{Cost Artemether/Lumefantrine treatment}
\usage{
cost_al(n_doses, cost_per_dose = 0.3)
}
\arguments{
\item{n_doses}{Number of tests}

\item{cost_per_dose}{Cost per dose is for a single dose (20/120 mg)}
}
\value{
AL costs
}
\description{
Note the cost per dose is for a single dose (20/120 mg). A treatment course typically
constitutes Artemether + lumefantrine given twice a day for 3 days following
weight-based guidelines:
\itemize{
  \item 5 to <15 kg: 20/120 mg
  \item 15 to <25 kg: 40/240 mg
  \item 25 to <35 kg: 60/360 mg
  \item >=35 kg: 80/480 mg
}
So course for a single adult (weighing >=35kg) may constitute
3 days x 2 times daily x 4 doses (4 x 20/120mg = 80/480mg) = 24 doses.
}
\references{
\strong{dosing information}

\url{https://www.who.int/publications/i/item/guidelines-for-malaria}

\strong{cost_per_dose}

Current default is the average cost for 20/120mg.

The Global Fund Pooled Procurement Mechanism Reference Pricing: Antimalarial medicines, version: quarter 1, 2022

\url{https://www.theglobalfund.org/en/sourcing-management/health-products/antimalarial-medicines/}.
}
