% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{cost_chloroquine}
\alias{cost_chloroquine}
\title{Cost Chloroquine treatment}
\usage{
cost_chloroquine(n_doses, cost_per_dose = 0.1/10, input_year = 2003, ...)
}
\arguments{
\item{n_doses}{Number of doses}

\item{cost_per_dose}{Cost per dose is for a single dose (250mg base each)}

\item{input_year}{Year the unit costs are reported in}

\item{...}{Additional arguments passed to `inflation_adjust()`}
}
\value{
Chloroquine costs
}
\description{
Cost Chloroquine treatment
}
\references{
\strong{cost_per_dose}

Assumes a full adult course is ~10 tablets of 250mg chloroquine base, and
costs $0.10 total

\url{https://www.msf.org/qa-act-now-get-malaria-treatment-works-africa}.
}
