% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bet_nets.R
\name{cost_dualai_itn}
\alias{cost_dualai_itn}
\title{Cost pyrethroid-chlorfenapyr (dual ai) ITN}
\usage{
cost_dualai_itn(
  n_dualai_itn,
  dualai_itn_unit_cost = 2.7,
  dualai_itn_delivery_cost = 1.5
)
}
\arguments{
\item{n_dualai_itn}{Number of pyrethroid-chlorfenapyr bed nets}

\item{dualai_itn_unit_cost}{Commodity unit cost per pyrethroid-chlorfenapyr ITN bed net.}

\item{dualai_itn_delivery_cost}{Cost to deliver one pyrethroid-chlorfenapyr ITN bet net.}
}
\value{
LLIN costs
}
\description{
Cost pyrethroid-chlorfenapyr (dual ai) ITN
}
\references{
\strong{dualai_itn_unit_cost}

Current default is the average cost for a pyrethroid-chlorfenapyr ITN, including,
  hooks, strings, bag and customisation.

The Global Fund Pooled Procurement Mechanism Reference Pricing: Insecticide-Treated Nets, accessed 16-12-2024

\url{https://www.theglobalfund.org/en/sourcing-management/health-products/long-lasting-insecticidal-nets/}.

\strong{dualai_itn_delivery_cost}

Sherrard-Smith et al (2022)

\url{https://www.thelancet.com/journals/lanplh/article/PIIS2542-5196(21)00296-5/fulltext}.
}
