% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{cost_inpatient}
\alias{cost_inpatient}
\title{Cost per inpatient visit}
\usage{
cost_inpatient(
  n_visits,
  cost_per_day,
  average_stay_duration = 3,
  input_year = 2021,
  ...
)
}
\arguments{
\item{n_visits}{Number of visits}

\item{cost_per_day}{Cost per day}

\item{average_stay_duration}{Average duration of stay, defaults to 3 days following Patouillard et al 2017.}

\item{input_year}{Year the unit costs are reported in}

\item{...}{Additional arguments passed to `inflation_adjust()`}
}
\value{
Inpatient costs
}
\description{
For information on country specific inpatient costs from WHO CHOICE see \code{?who_coice}
}
