% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{cost_inpatient}
\alias{cost_inpatient}
\title{Cost per inpatient visit}
\usage{
cost_inpatient(n_visits, cost_per_visit)
}
\arguments{
\item{n_visits}{Number of visits}

\item{cost_per_visit}{Cost per visit}
}
\value{
Inpatient costs
}
\description{
For information on country specific inpatient costs from WHO CHOICE see \code{?who_coice}
}
