% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iptp.R
\name{cost_iptp}
\alias{cost_iptp}
\title{Cost IPTp}
\usage{
cost_iptp(n_administrations, iptp_cost_per_administration = 0.79)
}
\arguments{
\item{n_administrations}{Number of IPTp doses}

\item{iptp_cost_per_administration}{Cost per dose delivered}
}
\value{
IPTp costs
}
\description{
Cost IPTp
}
\references{
\strong{Dosing information}

Starting as early as possible in the second trimester, IPTp-SP is
recommended for all pregnant women at each scheduled antenatal care (ANC)
visit until the time of delivery, provided that the doses are given at least
one month apart.

IPTp-SP should ideally be administered as directly observed therapy (DOT)
of three tablets sulfadoxine/pyrimethamine
(each tablet containing500 mg/25 mg SP) giving the total required dosage
of 1500 mg/75 mg SP.

\url{https://www.who.int/publications-detail-redirect/WHO-HTM-GMP-2014.4}

\strong{IPTp_cost_per_dose_delivered}

Current default is the total cost per administration of IPTp-SP
(health care worker time and commodity cost) with sulfadoxine–pyrimethamine.
Costs are in 2012 USD$.

Fernandes et al (2016). Table 2.

\url{https://malariajournal.biomedcentral.com/articles/10.1186/s12936-016-1539-4}
}
