% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{cost_microscopy}
\alias{cost_microscopy}
\title{Cost Microscopy}
\usage{
cost_microscopy(n_tests, cost_per_slide = 0.26, input_year = 2007, ...)
}
\arguments{
\item{n_tests}{Number of tests}

\item{cost_per_slide}{Cost per slide diagnostic performed}

\item{input_year}{Year the unit costs are reported in}

\item{...}{Additional arguments passed to `inflation_adjust()`}
}
\value{
Microscopy costs
}
\description{
Cost Microscopy
}
\references{
\strong{Microsopy_unit_cost}

Estimate of $0.26 per slide taken from Lubell et all (2007)

\url{https://pubmed.ncbi.nlm.nih.gov/18165484/}.
}
