% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{cost_microscopy}
\alias{cost_microscopy}
\title{Cost Microscopy}
\usage{
cost_microscopy(n_tests, cost_per_slide = 0.67)
}
\arguments{
\item{n_tests}{Number of tests performed}

\item{cost_per_slide}{Cost per slide diagnostic performed}
}
\value{
Microscopy costs
}
\description{
Cost Microscopy
}
\references{
\strong{Microsopy_unit_cost}

Estimate of $0.26 per slide taken from Lubell et all (2007)

`inflation_adjust(0.26, 2007, 2024)`

\url{https://pubmed.ncbi.nlm.nih.gov/18165484/}.
}
