% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{cost_outpatient}
\alias{cost_outpatient}
\title{Cost per outpatient visit}
\usage{
cost_outpatient(n_visits, cost_per_visit)
}
\arguments{
\item{n_visits}{Number of visits}

\item{cost_per_visit}{Cost per visit}
}
\value{
Outpatient costs
}
\description{
For information on country specific outpatient costs from WHO CHOICE see \code{?who_coice}
}
