% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acd.R
\name{cost_pacd}
\alias{cost_pacd}
\title{Cost pACD}
\usage{
cost_pacd(n_tested, cost_per_person_tested = 4.79, input_year = 2015, ...)
}
\arguments{
\item{n_tested}{Number of people tested}

\item{cost_per_person_tested}{Cost per person tested}

\item{input_year}{Year the unit costs are reported in}

\item{...}{Additional arguments passed to `inflation_adjust()`}
}
\value{
pACD costs
}
\description{
Cost for proactive case detection.
}
\references{
\strong{cost_per_person_tested}

Silumbe et al (2015)

\url{https://malariajournal.biomedcentral.com/articles/10.1186/s12936-015-0722-3}.
}
