% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bet_nets.R
\name{cost_pbo_itn}
\alias{cost_pbo_itn}
\title{Cost pyrethroid-PBO ITN}
\usage{
cost_pbo_itn(n_pbo_itn, pbo_itn_unit_cost = 3.51, pbo_itn_delivery_cost = 1.5)
}
\arguments{
\item{n_pbo_itn}{Number of pyrethroid-PBO bed nets}

\item{pbo_itn_unit_cost}{Commodity unit cost per pyrethroid-PBO ITN bed net.}

\item{pbo_itn_delivery_cost}{Cost to deliver one pyrethroid-PBO ITN bet net.}
}
\value{
LLIN costs
}
\description{
Cost pyrethroid-PBO ITN
}
\references{
\strong{pbo_itn_unit_cost}

Current default is the average cost for a pyrethroid-PBO ITN, including,
  hooks, strings, bag and customisation.

The Global Fund Pooled Procurement Mechanism Reference Pricing: Insecticide-Treated Nets, version: quarter 1, 2022

\url{https://www.theglobalfund.org/en/sourcing-management/health-products/long-lasting-insecticidal-nets/}.

\strong{pbo_itn_delivery_cost}

Sherrard-Smith et al (2022)

\url{https://www.thelancet.com/journals/lanplh/article/PIIS2542-5196(21)00296-5/fulltext}.
}
