% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccine.R
\name{cost_r21}
\alias{cost_r21}
\title{Cost R21}
\usage{
cost_r21(
  n_doses,
  r21_cost_per_dose = 4,
  r21_consumables_cost = 1.52,
  r21_delivery_cost = 1.48
)
}
\arguments{
\item{n_doses}{Number of R21 doses}

\item{r21_cost_per_dose}{Cost per R21 dose}

\item{r21_consumables_cost}{Cost for consumables for one dose (e.g injection and reconstitution syringes, safety box etc.)}

\item{r21_delivery_cost}{Cost for delivery of one dose}
}
\value{
R21 costs
}
\description{
Cost R21
}
\references{
\strong{r21_cost_per_dose}

Current default is based on the $4 per dose as used in
\url{https://www.thelancet.com/journals/laninf/article/PIIS1473-3099(23)00816-2/fulltext}

Penny et al (2016)
\url{https://www.thelancet.com/journals/lancet/article/PIIS0140-6736(15)00725-4/fulltext}.

\strong{r21_consumables_cost}

Penny et al (2016)
\url{https://www.thelancet.com/journals/lancet/article/PIIS0140-6736(15)00725-4/fulltext}.

\strong{r21_delivery_cost}

Following methodology summarising available information from trials and MVIP in SI of
\url{https://www.thelancet.com/journals/laninf/article/PIIS1473-3099(23)00816-2/fulltext}

Delivery cost per dose:
Age-based: $1.48 (default)
Seasonal: $3.75
Hybrid: $2.36
}
