% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acd.R
\name{cost_racd}
\alias{cost_racd}
\title{Cost rACD}
\usage{
cost_racd(n_tested, cost_per_person_tested = 38.63)
}
\arguments{
\item{n_tested}{Number of people tested}

\item{cost_per_person_tested}{Cost per person tested}
}
\value{
pACD costs
}
\description{
Cost for reactive case detection.
}
\references{
\strong{cost_per_person_tested}

Larson et al (2016)

\url{https://malariajournal.biomedcentral.com/articles/10.1186/s12936-016-1457-5}.
}
