% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{cost_rdt}
\alias{cost_rdt}
\title{Cost RDTs}
\usage{
cost_rdt(
  n_tests,
  rdt_unit_cost = 0.46,
  delivery_mark_up = 0.15,
  input_year = 2022,
  ...
)
}
\arguments{
\item{n_tests}{Number of tests}

\item{rdt_unit_cost}{Unit cost for rapid diagnostic test}

\item{delivery_mark_up}{A mark up for in-country delivery to a public health facility.
Expressed as a proportion of the test unit cost.}

\item{input_year}{Year the unit costs are reported in}

\item{...}{Additional arguments passed to `inflation_adjust()`}
}
\value{
RDT costs
}
\description{
RDTs are used for diagnosis of malaria. When costing it is also common to add
additional costs for RDTs used to diagnose non-malaria fevers.
}
\references{
\strong{rdt_unit_cost}

Current default is the average cost for an RDT from Normal orders, Reference price per test EXW.

The Global Fund Pooled Procurement Mechanism Reference Pricing: RDTs, version: quarter 1, 2022

\url{https://www.theglobalfund.org/en/sourcing-management/health-products/hiv-malaria-rapid-diagnostic-tests/}.

\strong{delivery_mark_up}

Assummed 15%

Patouillard et al (2017)

\url{https://gh.bmj.com/content/2/2/e000176}
}
