% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccine.R
\name{cost_rtss}
\alias{cost_rtss}
\title{Cost RTS,S}
\usage{
cost_rtss(
  n_doses,
  rtss_cost_per_dose = 10.02,
  rtss_consumables_cost = 1.52,
  rtss_delivery_cost = 1.48
)
}
\arguments{
\item{n_doses}{Number of RTS,S doses}

\item{rtss_cost_per_dose}{Cost per RTS,S dose}

\item{rtss_consumables_cost}{Cost for consumables for one dose (e.g injection and reconstitution syringes, safety box etc.)}

\item{rtss_delivery_cost}{Cost for delivery of one dose}
}
\value{
RTS,S costs
}
\description{
Cost RTS,S
}
\references{
\strong{rtss_cost_per_dose}

Current default is based on the EUR9.30 per dose quoted in
\url{https://www.unicef.org/supply/media/19456/file/Malaria\%20-\%20Vaccine\%20-\%20QA\%20-\%20October\%202023\%20-\%20English\%20.pdf}
}
