% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc.R
\name{cost_smc}
\alias{cost_smc}
\title{Cost SMC}
\usage{
cost_smc(n_doses, smc_cost_per_dose_delivered = 0.9075, input_year = 2016, ...)
}
\arguments{
\item{n_doses}{Number of SMC doses. Numeric scalar or vector.}

\item{smc_cost_per_dose_delivered}{Cost per dose delivered. Numeric scalar or vector.}

\item{input_year}{Year the unit costs are reported in}

\item{...}{Additional arguments passed to `inflation_adjust()`}
}
\value{
SMC costs
}
\description{
Cost SMC
}
\references{
\strong{smc_cost_per_dose_delivered}

Current default is the average economic cost of administering four monthly
  SMC cycles of sulfadoxine–pyrimethamine plus amodiaquine for children
  younger than 5 years across seven countries in the Sahel subregion
  (Burkina Faso, Chad, Guinea, Mali, Niger, Nigeria, and The Gambia)
  in 2016. These include multiple delivery methods (door to door, fixed point,
   mobile point, mixed). The cost per dose delivered is the average cost of
   four monthly SMC cycles divided by the number of cycles (3.63 / 4). Costs
   are in 2016 USD$.

Gilmartin et al (2021)

\url{https://www.thelancet.com/journals/langlo/article/PIIS2214-109X(20)30475-7/fulltext}.
}
