% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpi}
\alias{cpi}
\title{World Bank Consumer Price Index (CPI)}
\format{
A data frame with 1816 rows and 10 variables:
\describe{
  \item{year}{Year}
  \item{region}{Region}
  \item{cpi}{Consumer Price Index}
}
}
\source{
\url{https://data.worldbank.org/indicator/FP.CPI.TOTL.ZG?locations=ZG}
}
\usage{
cpi
}
\description{
Regional CPI by year, taken as median of country-level CPI
}
\details{
Data from: World Bank.
}
\keyword{datasets}
