% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflate.R
\name{inflation_adjust}
\alias{inflation_adjust}
\title{Inflate cost to a target year using SSA CPI}
\usage{
inflation_adjust(cost, cost_year, target_year, region = "Sub-Saharan Africa")
}
\arguments{
\item{cost}{Numeric value of the original cost.}

\item{cost_year}{Integer indicating the year corresponding to `cost`.}

\item{target_year}{Integer indicating the year to adjust the cost to.}

\item{region}{World region, defaults to SSA}
}
\value{
Numeric value of the inflation-adjusted cost in `target_year` dollars.
}
\description{
Adjust a given cost from its original year to a specified target year using
the Consumer Price Index (CPI) values from the `cpi` dataset. Note this appoach
uses region median CPIs estimates. For a more correct approach at the country
level see \url{https://linkinghub.elsevier.com/retrieve/pii/S1098-3015(19)32149-7}
}
\details{
The `ssa_cpi` dataset must be included in the package and contain columns:
- `year`: Calendar year.
- `cpi`: CPI index value for that year.
}
\examples{
# Adjust $0.26 from 2007 to 2024
inflation_adjust(0.26, 2007, 2024)

}
