% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflate.R
\name{set_target_year}
\alias{set_target_year}
\title{Set the global target year for inflation adjustments}
\usage{
set_target_year(year)
}
\arguments{
\item{year}{Target year to use when adjusting for inflation}
}
\description{
Set the global target year for inflation adjustments
}
